Moved histogram series into a subfolder.
